Add ability to update [Val], [Updated_Date], and [Update_User_ID] to SavePerIndDataApi
Also will return error if user doesn't exists in the database, or the
user exists, but is not active. This commit, does not check if User has
correct Category permissions, nor have data validation. In the case of
Category permissions, currently any authenticated user can edit any of
the record by POSTING to the api. And in the case of data validation, it
will simply return error to the front end saying it could not convert
successfully the stirng into a float (Maye this is all that's needed,
but will probably fix in a later commit for a more graceful response).