Fix missing pms info on FleetDataCollection's JS DataTable
Was calling the wrong api, which returned a list of PMS info that the client
has WU permissions to. This pms list is not the entire PMS list, so some
records on the JS DataTable is missing emp info even though its pms
column is populated.

To fix it, it is now calling the correct api.