Add DEV or QA flag to PMU_DjangoWebApps/templates/template.base.html
This will make clear on all webpages that the current server is either
Dev, QA or Prod. If it is Dev or QA, it will output a banner at the top
of all pages saying so. If it is Prod, it will have no banner.

Set the variable SERVER_TYPE in secret_settings.py to either 'Dev',
'QA', or 'Prod' to specify what type of server is the hosting machine.

If the SERVER_TYPE variable's value is not recognize, it will output a banner
on all webpages saying so.